requirements.txt
# Pyrogram y sus dependencias de rendimiento
Pyrogram>=2.0.106
Tgcrypto

# Interacción con MongoDB (asíncrona)
motor>=3.1.1

# Carga de variables de entorno
python-dotenv>=1.0.0

# Descargador principal de video/audio
yt-dlp>=2023.7.6

# Para la API de Spotify (búsqueda)
spotipy>=2.23.0

# Dependencia para descargas genéricas (ej. carátulas)
requests>=2.31.0

# Para extraer metadatos de audio si es necesario
Pillow